bugfix: removed warnings for linux ppl

+++ NaoTHSoccer/Source/Tools/DataStructures/Spline.cpp
@@ -32,7 +32,7 @@ void band_matrix::resize(int dim, int n_u, int n_l)
 int band_matrix::dim() const
 {
     if(m_upper.size()>0) {
-        return m_upper[0].size();
+        return static_cast<int>(m_upper[0].size());
     } else {
         return 0;
     }
@@ -176,7 +176,7 @@ void spline::set_points(const std::vector<double>& x,
     assert(x.size()>2);
     m_x=x;
     m_y=y;
-    int   n=x.size();
+    int   n = static_cast<int>(x.size());
     // TODO: maybe sort x and y, rather than returning an error
     for(int i=0; i<n-1; i++) {
         assert(m_x[i]<m_x[i+1]);
